Refactor admin dashboard and layout: Remove Hadeeths link, update font size input to accept rem values, and enhance styling for announcements and prayer times. Implement live update functionality for settings and adjust routes to remove HadeethController.

+++ app/Models/SlidingText.php
@@ -20,7 +20,7 @@ class SlidingText extends Model
     protected $casts = [
         'is_active' => 'boolean',
         'animation_speed' => 'integer',
-        'font_size' => 'integer',
+        'font_size' => 'float', // Updated to allow decimal values for rem
         'display_order' => 'integer'
     ];
